+++ core/.circleci/check-markdown-links.sh
@@ -1,3 +1,8 @@
+#
+# Copyright (c) 2021 - present Kurtosis Technologies Inc.
+# All Rights Reserved.
+#
+
 set -euo pipefail
 script_dirpath="$(cd "$(dirname "${0}")" && pwd)"
 root_dirpath="$(dirname "${script_dirpath}")"