+++ core/api_container/server/service_network/user_service_launcher/files_artifact_expander/files_artifact_expander.go
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * Copyright (c) 2021 - present Kurtosis Technologies Inc.
  * All Rights Reserved.
  */
 
@@ -7,18 +7,14 @@ package files_artifact_expander
 
 import (
 	"context"
-	"fmt"
 	"github.com/docker/go-connections/nat"
-	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/container_name_provider"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/service_network_types"
 	"github.com/kurtosis-tech/kurtosis/commons"
 	"github.com/kurtosis-tech/kurtosis/commons/docker_manager"
 	"github.com/kurtosis-tech/kurtosis/commons/enclave_data_volume"
-	"github.com/kurtosis-tech/kurtosis/commons/volume_naming_consts"
+	"github.com/kurtosis-tech/kurtosis/commons/object_name_providers"
 	"github.com/palantir/stacktrace"
 	"path"
-	"strings"
-	"time"
 )
 
 const (
@@ -41,13 +37,11 @@ Class responsible for taking an artifact containing compressed files and uncompr
 	into a Docker volume that will be mounted on a new service
  */
 type FilesArtifactExpander struct {
-	enclaveNameElems []string
-
 	enclaveDataVolName string
 
 	dockerManager *docker_manager.DockerManager
 
-	containerNameElemsProvider *container_name_provider.ContainerNameElementsProvider
+	enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider
 
 	testNetworkId string
 
@@ -56,8 +50,8 @@ type FilesArtifactExpander struct {
 	filesArtifactCache *enclave_data_volume.FilesArtifactCache
 }
 
-func NewFilesArtifactExpander(enclaveNameElems []string, enclaveDataVolName string, dockerManager *docker_manager.DockerManager, containerNameElemsProvider *container_name_provider.ContainerNameElementsProvider, testNetworkId string, freeIpAddrTracker *commons.FreeIpAddrTracker, filesArtifactCache *enclave_data_volume.FilesArtifactCache) *FilesArtifactExpander {
-	return &FilesArtifactExpander{enclaveNameElems: enclaveNameElems, enclaveDataVolName: enclaveDataVolName, dockerManager: dockerManager, containerNameElemsProvider: containerNameElemsProvider, testNetworkId: testNetworkId, freeIpAddrTracker: freeIpAddrTracker, filesArtifactCache: filesArtifactCache}
+func NewFilesArtifactExpander(enclaveDataVolName string, dockerManager *docker_manager.DockerManager, enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider, testNetworkId string, freeIpAddrTracker *commons.FreeIpAddrTracker, filesArtifactCache *enclave_data_volume.FilesArtifactCache) *FilesArtifactExpander {
+	return &FilesArtifactExpander{enclaveDataVolName: enclaveDataVolName, dockerManager: dockerManager, enclaveObjNameProvider: enclaveObjNameProvider, testNetworkId: testNetworkId, freeIpAddrTracker: freeIpAddrTracker, filesArtifactCache: filesArtifactCache}
 }
 
 func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
@@ -66,7 +60,7 @@ func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
 		artifactIdsToExpand map[string]bool) (map[string]string, error) {
 	artifactIdsToVolNames := map[string]string{}
 	for artifactId := range artifactIdsToExpand {
-		destVolName := expander.getExpandedFilesArtifactVolName(serviceId, artifactId)
+		destVolName := expander.enclaveObjNameProvider.ForFilesArtifactExpansionVolume(string(serviceId), artifactId)
 		artifactIdsToVolNames[artifactId] = destVolName
 	}
 
@@ -94,8 +88,8 @@ func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
 			destVolName:                  destVolMntDirpathOnExpander,
 		}
 
-		containerNameElems := expander.containerNameElemsProvider.GetForFilesArtifactExpander(serviceId, artifactId)
-		if err := expander.runExpanderContainer(ctx, containerNameElems, containerCmd, volumeMounts); err != nil {
+		containerName := expander.enclaveObjNameProvider.ForFilesArtifactExpanderContainer(serviceId, artifactId)
+		if err := expander.runExpanderContainer(ctx, containerName, containerCmd, volumeMounts); err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred running the expander container")
 		}
 	}
@@ -106,26 +100,11 @@ func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
 // ====================================================================================================
 //                                       Private Helper Functions
 // ====================================================================================================
-func (expander *FilesArtifactExpander) getExpandedFilesArtifactVolName(
-	serviceId service_network_types.ServiceID,
-	artifactId string) string {
-	timestampStr := time.Now().Format(volume_naming_consts.GoTimestampFormat)
-	prefix := strings.Join(expander.enclaveNameElems, "_")
-	// TODO Standardize this!
-	return fmt.Sprintf(
-		"%v_%v_%v_%v",
-		timestampStr,
-		prefix,
-		serviceId,
-		artifactId,
-	)
-}
-
 // NOTE: This is a separate function so we can defer the releasing of the IP address and guarantee that it always
 //  goes back into the IP pool
 func (expander *FilesArtifactExpander) runExpanderContainer(
 		ctx context.Context,
-		containerNameElems []string,
+		containerName string,
 		containerCmd []string,
 		volumeMounts map[string]string) error {
 	// NOTE: This silently (temporarily) uses up one of the user's requested IP addresses with a container
@@ -138,15 +117,16 @@ func (expander *FilesArtifactExpander) runExpanderContainer(
 	}
 	defer expander.freeIpAddrTracker.ReleaseIpAddr(containerIp)
 
-	containerId, err := expander.dockerManager.CreateAndStartContainer(
+	containerId, _, err := expander.dockerManager.CreateAndStartContainer(
 		ctx,
 		dockerImage,
-		containerNameElems,
+		containerName,
 		expander.testNetworkId,
 		containerIp,
 		map[docker_manager.ContainerCapability]bool{},
 		docker_manager.DefaultNetworkMode,
-		map[nat.Port]*nat.PortBinding{},
+		map[nat.Port]bool{},
+		false, // We don't need to publish any ports for the files artifact-expanding image
 		nil, // No ENTRYPOINT overriding needed
 		containerCmd,
 		map[string]string{}, // No env variables