+++ core/commons/volume_naming_consts/volume_naming_consts.go
@@ -1,11 +1,9 @@
 /*
- * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * Copyright (c) 2021 - present Kurtosis Technologies Inc.
  * All Rights Reserved.
  */
 
 package volume_naming_consts
 
 const (
-	// These should represent the same format of YYYY-MM-DDTHH.mm.ss
-	GoTimestampFormat = "2006-01-02T15.04.05"
 )