+++ core/initializer/auth/access_controller/device_auth_access_controller_test.go
@@ -1,4 +1,5 @@
-/* * Copyright (c) 2020 - present Kurtosis Technologies LLC.
+/*
+ * Copyright (c) 2021 - present Kurtosis Technologies Inc.
  * All Rights Reserved.
  */