+++ core/initializer/test_suite_launcher/test_suite_launcher.go
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * Copyright (c) 2021 - present Kurtosis Technologies Inc.
  * All Rights Reserved.
  */
 
@@ -12,8 +12,8 @@ import (
 	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_consts"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/kurtosis_testsuite_docker_api"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/kurtosis_testsuite_rpc_api_consts"
-	"github.com/kurtosis-tech/kurtosis/api_container/server/optional_host_port_binding_supplier"
 	"github.com/kurtosis-tech/kurtosis/commons/docker_manager"
+	"github.com/kurtosis-tech/kurtosis/commons/object_name_providers"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"net"
@@ -27,15 +27,13 @@ const (
 	// will be told to listen on
 	portForDebuggersRunningOnTestsuite     = 2778
 	protocolForDebuggersRunningOnTestsuite = "tcp"
-
-	// Identifier included in hostnames of containers running for metadata acquisition
-	metadataAcquisitionContainerNameLabel = "metadata-acquisition"
-
-	testsuiteContainerNameSuffix = "testsuite"
 )
 
+var testsuiteRpcPort = nat.Port(fmt.Sprintf("%v/%v", kurtosis_testsuite_rpc_api_consts.ListenPort, kurtosis_testsuite_rpc_api_consts.ListenProtocol))
+var testsuiteDebuggerPort = nat.Port(fmt.Sprintf("%v/%v", portForDebuggersRunningOnTestsuite, protocolForDebuggersRunningOnTestsuite))
+
 type TestsuiteContainerLauncher struct {
-	executionInstanceUuid string
+	testsuiteExObjNameProvider *object_name_providers.TestsuiteExecutionObjectNameProvider
 
 	testsuiteImage string
 
@@ -45,11 +43,11 @@ type TestsuiteContainerLauncher struct {
 	// The JSON-serialized custom params object that will be passed as-is to the testsuite
 	customParamsJson string
 
-	optionalHostPortBindingSupplier *optional_host_port_binding_supplier.OptionalHostPortBindingSupplier
+	shouldPublishPorts bool
 }
 
-func NewTestsuiteContainerLauncher(executionInstanceUuid string, testsuiteImage string, suiteLogLevel string, customParamsJson string, optionalHostPortBindingSupplier *optional_host_port_binding_supplier.OptionalHostPortBindingSupplier) *TestsuiteContainerLauncher {
-	return &TestsuiteContainerLauncher{executionInstanceUuid: executionInstanceUuid, testsuiteImage: testsuiteImage, suiteLogLevel: suiteLogLevel, customParamsJson: customParamsJson, optionalHostPortBindingSupplier: optionalHostPortBindingSupplier}
+func NewTestsuiteContainerLauncher(testsuiteExObjNameProvider *object_name_providers.TestsuiteExecutionObjectNameProvider, testsuiteImage string, suiteLogLevel string, customParamsJson string, shouldPublishPorts bool) *TestsuiteContainerLauncher {
+	return &TestsuiteContainerLauncher{testsuiteExObjNameProvider: testsuiteExObjNameProvider, testsuiteImage: testsuiteImage, suiteLogLevel: suiteLogLevel, customParamsJson: customParamsJson, shouldPublishPorts: shouldPublishPorts}
 }
 
 /*
@@ -83,15 +81,11 @@ func (launcher TestsuiteContainerLauncher) LaunchMetadataAcquiringContainer(
 
 	suiteContainerDesc := "metadata-providing testsuite container"
 	log.Infof("Launching %v...", suiteContainerDesc)
-	testsuiteContainerNameElems := []string{
-		launcher.executionInstanceUuid,
-		metadataAcquisitionContainerNameLabel,
-		testsuiteContainerNameSuffix,
-	}
+	containerName := launcher.testsuiteExObjNameProvider.ForMetadataAcquiringTestsuiteContainer()
 	testsuiteContainerId, debuggerPortHostBinding, err := launcher.createAndStartTestsuiteContainerWithDebuggingPortIfNecessary(
 		ctx,
 		dockerManager,
-		testsuiteContainerNameElems,
+		containerName,
 		bridgeNetworkId,
 		nil,   // Nil because the bridge network will assign IPs on its own (and won't know what IPs are already used)
 		testsuiteEnvVars,
@@ -144,18 +138,15 @@ func (launcher TestsuiteContainerLauncher) LaunchTestRunningContainer(
 
 	suiteContainerDesc := "test-running testsuite container"
 	log.Infof("Launching %v....", suiteContainerDesc)
-	suiteContainerNameElems := []string{
-		launcher.executionInstanceUuid,
-		testName,
-		testsuiteContainerNameSuffix,
-	}
+	_, enclaveObjNameProvider := launcher.testsuiteExObjNameProvider.ForTestEnclave(testName)
+	containerName := enclaveObjNameProvider.ForTestRunningTestsuiteContainer()
 	volumeMountpoints := map[string]string{
 		enclaveDataVolName: kurtosis_testsuite_docker_api.EnclaveDataVolumeMountpoint,
 	}
-	suiteContainerId, debuggerPortHostBinding, err := launcher.createAndStartTestsuiteContainerWithDebuggingPortIfNecessary(
+	suiteContainerId, hostPortBindings, err := launcher.createAndStartTestsuiteContainerWithDebuggingPortIfNecessary(
 		ctx,
 		dockerManager,
-		suiteContainerNameElems,
+		containerName,
 		networkId,
 		testsuiteContainerIp,
 		testSuiteEnvVars,
@@ -171,7 +162,7 @@ func (launcher TestsuiteContainerLauncher) LaunchTestRunningContainer(
 		suiteContainerId,
 		func() bool { return functionCompletedSuccessfully },
 	)
-	logSuccessfulSuiteContainerLaunch(log, suiteContainerDesc, debuggerPortHostBinding)
+	logSuccessfulSuiteContainerLaunch(log, suiteContainerDesc, hostPortBindings)
 
 	functionCompletedSuccessfully = true
 	return suiteContainerId, nil
@@ -184,57 +175,29 @@ func (launcher TestsuiteContainerLauncher) LaunchTestRunningContainer(
 func (launcher TestsuiteContainerLauncher) createAndStartTestsuiteContainerWithDebuggingPortIfNecessary(
 		ctx context.Context,
 		dockerManager *docker_manager.DockerManager,
-		containerNameElems []string,
+		name string,
 		networkId string,
 		containerIpAddr net.IP,
 		envVars map[string]string,
 		volumeMountpoints map[string]string,
-	) (string, *nat.PortBinding, error) {
+	) (string, map[nat.Port]*nat.PortBinding, error) {
 
-	testsuiteRpcPort, err := nat.NewPort(kurtosis_testsuite_rpc_api_consts.ListenProtocol, strconv.Itoa(kurtosis_testsuite_rpc_api_consts.ListenPort))
-	if err != nil {
-		return "", nil, stacktrace.Propagate(
-			err,
-			"An error occurred creating the port object on '%v/%v' to represent the testsuite RPC API",
-			kurtosis_testsuite_rpc_api_consts.ListenProtocol,
-			kurtosis_testsuite_rpc_api_consts.ListenPort,
-		)
-	}
-
-	testsuiteDebuggerPort, err := nat.NewPort(protocolForDebuggersRunningOnTestsuite, strconv.Itoa(portForDebuggersRunningOnTestsuite))
-	if err != nil {
-		return "", nil, stacktrace.Propagate(
-			err,
-			"An error occurred creating the port object on '%v/%v' to represent the testsuite debugger port",
-			protocolForDebuggersRunningOnTestsuite,
-			portForDebuggersRunningOnTestsuite,
-		)
-	}
 
 	usedPorts := map[nat.Port]bool{
 		testsuiteRpcPort: true,
 		testsuiteDebuggerPort: true,
 	}
 
-	usedPortsWithHostBindings, err := launcher.optionalHostPortBindingSupplier.BindPortsToHostIfNeeded(usedPorts)
-	if err != nil {
-		return "", nil, stacktrace.Propagate(err, "An error occurred binding the test suite's used ports to host port bindings where necessary")
-	}
-
-	debuggerHostPortBinding, found := usedPortsWithHostBindings[testsuiteDebuggerPort]
-	if !found {
-		return "", nil, stacktrace.NewError("Couldn't find an entry for the debugger port even though it should be in the map!")
-	}
-
-	containerId, err := dockerManager.CreateAndStartContainer(
+	containerId, hostPortBindings, err := dockerManager.CreateAndStartContainer(
 		ctx,
 		launcher.testsuiteImage,
-		containerNameElems,
+		name,
 		networkId,
 		containerIpAddr,
 		map[docker_manager.ContainerCapability]bool{}, 	// No extra capabilities needed for testsuite containers
 		docker_manager.DefaultNetworkMode,  			// No special networking modes for testsuite containers
-		usedPortsWithHostBindings,
+		usedPorts,
+		launcher.shouldPublishPorts,
 		nil, // Nil ENTRYPOINT args because we expect the test suite image to be parameterized with variables
 		nil, // Nil CMD args because we expect the test suite image to be parameterized with variables
 		envVars,
@@ -246,19 +209,20 @@ func (launcher TestsuiteContainerLauncher) createAndStartTestsuiteContainerWithD
 		return "", nil, stacktrace.Propagate(err, "An error occurred creating and starting the testsuite container")
 	}
 
-	return containerId, debuggerHostPortBinding, nil
+	return containerId, hostPortBindings, nil
 }
 
 func logSuccessfulSuiteContainerLaunch(
 		log *logrus.Logger,
 		suiteContainerDesc string,
-		debuggerPortHostBinding *nat.PortBinding) {
+		hostPortBindings map[nat.Port]*nat.PortBinding) {
 	suiteLaunchSupplementalLogInfo := ""
-	if debuggerPortHostBinding != nil {
+	debuggerHostPortBinding, found := hostPortBindings[testsuiteDebuggerPort]
+	if found {
 		suiteLaunchSupplementalLogInfo = fmt.Sprintf(
 			" with debugger port bound to host port %v:%v (if a debugger is running, you may need to connect to this port to proceed)",
-			debuggerPortHostBinding.HostIP,
-			debuggerPortHostBinding.HostPort,
+			debuggerHostPortBinding.HostIP,
+			debuggerHostPortBinding.HostPort,
 		)
 	}
 	log.Infof("Successfully created %v%v", suiteContainerDesc, suiteLaunchSupplementalLogInfo)