+++ core/initializer/test_suite_runner/test_suite_runner_test.go
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * Copyright (c) 2021 - present Kurtosis Technologies Inc.
  * All Rights Reserved.
  */
 
@@ -18,7 +18,7 @@ func TestBlockedExecutionWhenNoPerms(t *testing.T) {
 	perms := permissions.FromPermissionsSet(map[string]bool{})
 	result, err := RunTests(
 		perms,
-		"1234-abcd",
+		nil,
 		"5678-efgh",
 		nil,
 		nil,
@@ -38,7 +38,7 @@ func TestBlockedExecutionWhenRestrictedPerms(t *testing.T) {
 	})
 	result, err := RunTests(
 		perms,
-		"1234-abcd",
+		nil,
 		"5678-efgh",
 		nil,
 		nil,